+++ src/commands/config-view.ts
@@ -32,7 +32,11 @@ class ConfigViewCommand extends Command {
 	protected async exec(): Promise<string> {
 		this.log.info('config', this.datadir.configPath);
 
-		return this.datadir.configToml;
+		if (this.args.options.json) {
+			return JSON.stringify(this.datadir.config);
+		} else {
+			return this.datadir.configToml;
+		}
 	}
 }