+++ src/commands/poa-evictee-vote.ts
@@ -233,7 +233,15 @@ class POAEvictionVoteCommand extends Command<Args> {
 
 		this.stopSpinner();
 
-		return message;
+		if (this.args.options.json) {
+			return JSON.stringify({
+				voter: this.account.address,
+				evictee: this.args.address,
+				verdict: this.args.options.verdict
+			});
+		} else {
+			return message;
+		}
 	}
 }