+++ CHANGELOG.md
@@ -5,11 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Changed
+
+-   All commands now by default output formatted data unless specified by the `-j, --json` flag. (Better scripting experience)
+
 ## v1.3.0
 
 ### Added
 
 -   POA eviction commands
+    -   `poa evictee new`
+    -   `poa evictee list`
+    -   `poa evictee vote`
+
+### Changed
+
+-   Some commands have been renamed for better command structure
+    -   `poa nominate` -> `poa nominee new`
+    -   `poa vote` -> `poa nominee vote`
+    -   `poa nomineelist` -> `poa nominee list`
 
 ## v1.2.2